# src/main/resources/swagger/api_cuentas_banca_v1.yml
openapi: 3.0.3
info:
  title: Swagger Bank Management - OpenAPI 3.0
  description: |-
    This is a bank customer account management server, where the customer can use and consult their banking products.

  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://managementBank.swagger.io/api/v3
...
paths:
...
  /report/balanceAllProduct/{numberDocument}:
    get:
      tags:
        - reports
      summary: Gets report the balance of all products from client
      description: Gets report the balance of all products from client
      operationId: getReportBalanceProducts
      parameters:
        - name: numberDocument
          in: path
          description: client document number
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: client not found
  /report/commissionbyProduct/{accountNumber}:
    get:
      tags:
        - reports
      summary: Gets report commission by products from client
      description: Gets report commission by products from client
      operationId: getReportCommisionByProduct
      parameters:
        - name: accountNumber
          in: path
          description: client document number
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: client not found
  /report/movementsCards/{numberDocument}:
    get:
      tags:
        - reports
      summary: Gets report movements from debit and credit card
      description: Gets report movements from debit and credit card
      operationId: getReportMovementsCards
      parameters:
        - name: numberDocument
          in: path
          description: client document number
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: client not found
...